restructuring validation code
